added filingnumber in case details search
